requirements_hf.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.0
shap==0.43.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psutil==5.9.6
joblib==1.3.2